# documentation/dev/apiv2/swagger/prompb/rpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "prompb/rpc.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v2/query": {
      "get": {
        "operationId": "Instant",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/prometheusQueryResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "range.min_time",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "range.max_time",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "expr",
            "description": "PromQL formatted query expression.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Query"
        ]
      }
    },
...
  },
  "definitions": {
...
    "prometheusLabel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
...
    "prometheusQueryResponse": {
      "type": "object",
      "properties": {
        "series": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/prometheusTimeSeries"
          }
        }
      }
    },
...
    "prometheusSample": {
      "type": "object",
      "properties": {
        "value": {
          "type": "number",
          "format": "double"
        },
        "timestamp_ms": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "prometheusTimeSeries": {
      "type": "object",
      "properties": {
        "labels": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/prometheusLabel"
          }
        },
        "samples": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/prometheusSample"
          }
        }
      }
    },
...
  }
}